requirements.txt
Flask==2.3.2
pytesseract==0.3.13
pillow>=10.0.0
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
requests==2.31.0